requirements.txt
aiohttp==3.6.2
aiorpcX==0.18.3
async-timeout==3.0.1
attrs==19.3.0
certifi==2019.9.11
chardet==3.0.4
coinbase==2.1.0
defusedxml==0.6.0
Django==2.2.7
django-allauth==0.40.0
django-crispy-forms==1.8.1
electrumX==1.13.0
idna==2.8
multidict==4.5.2
oauthlib==3.1.0
pipenv==2018.11.26
plyvel==1.1.0
pycryptodome==3.9.3
pylru==1.2.0
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
virtualenv==16.7.7
virtualenv-clone==0.5.3
websockets==8.1
yarl==1.3.0
